Extra promise to preserve unhandled rejection
In case you care about unhandled rejections, the previous version of
this code would eat one in some situations.

The `result` promise always got `catch` called on it, so a rejection on
that promise could never be considered "unhandled".

Calling `.then(value => value)` creates a new promise that will resolve
or reject the same way as `result`. If this new promise rejects and it
has never been used for anything, it will be "unhandled".